Being stupid and fixing the ridiculously long running

+++ A0C2019/Models/Responses/2019/Day3Response.cs
@@ -7,7 +7,7 @@ namespace AdventOfCode.Models.Responses._2019
 {
     public class Day3Response : BaseResponse
     {
-        public Day3Response(double answer1, int answer2)
+        public Day3Response(double answer1, double answer2)
         {
             this.Answer1 = answer1.ToString();
             this.Answer2 = answer2.ToString();
